refactor(init): extract action-target resolution (#519)
